+++ NEWS.md
@@ -10,6 +10,7 @@
 * `groups()` and `as.data.table()` methods for `featureGroups`: optionally consider feature areas instead of peak intensities.
 * `plotSilhouettes()` method for `compoundsCluster`
 * Added `rGroups` argument to subset operator for `featureGroups` to subset by replicate groups (equivalent to `rGroups` argument to `filter()`).
+* Improved logging of output from CLI tools (e.g. OpenMS, MetFrag, SIRUS, ...) 
 
 ## May 2019